validate phone numbers

+++ src/main/java/com/example/RegisterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RegisterRepository extends JpaRepository<RegisterEntity,Integer > {
+public interface RegisterRepository extends JpaRepository<RegisterEntity,Long > {
 }
